refactor : `PathService` 경로조회 로직에 요금계산 코드 추가
