+++ src/main/java/nextstep/subway/line/application/PathSearch.java
@@ -1,7 +1,7 @@
 package nextstep.subway.line.application;
 
 import nextstep.subway.line.domain.Path;
-import nextstep.subway.line.dto.path.PathResult;
+import nextstep.subway.line.domain.PathResult;
 
 public interface PathSearch {
